+++ demo/src/main/java/com/telusko/demo/repo/ProjectRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ProjectRepository extends CrudRepository<Project, Long> {
+public interface ProjectRepository extends CrudRepository<Project, Integer> {
     Optional<Project> findById(int id);
 
 }
